requirements.txt
Django==4.1.3
langflow
drf_yasg
whitenoise
django-cors-headers==4.3.1
djangorestframework==3.14.0